fix: mv example
